fix: robustness and cleanup across MATLAB and Python
MATLAB:
- Remove dead DEBUG fprintf blocks from fastlpr_kdf.m
- Use mfilename-relative paths in generate_lwp_formulas.m

Python:
- Add grid shape validation in convolution.py
- Chunked eval for large LWP formulas (Python parser limit)
- Disable fastmath in Numba JIT for numerical precision
- Guard against h=0 division in heat_kernel
- Fix complex dtype mismatch in spreading_fused.py

+++ fastLPR_py/src/fastlpr/spreading_fused.py
@@ -240,7 +240,7 @@ if HAS_NUMBA:
                 hx.astype(np.float64), tau.astype(np.float64),
                 strides.astype(np.int64)
             )
-            Ftau_flat[:] = Ftau_real + 1j * Ftau_imag
+            Ftau_flat[:] = (Ftau_real + 1j * Ftau_imag).astype(Ftau_flat.dtype)
         else:
             fused_spreading_1d_real(
                 xmod.astype(np.float64),
@@ -265,7 +265,7 @@ if HAS_NUMBA:
                 hx.astype(np.float64), tau.astype(np.float64),
                 strides.astype(np.int64)
             )
-            Ftau_flat[:] = Ftau_real + 1j * Ftau_imag
+            Ftau_flat[:] = (Ftau_real + 1j * Ftau_imag).astype(Ftau_flat.dtype)
         else:
             fused_spreading_2d_real(
                 xmod.astype(np.float64),
@@ -290,7 +290,7 @@ if HAS_NUMBA:
                 hx.astype(np.float64), tau.astype(np.float64),
                 strides.astype(np.int64)
             )
-            Ftau_flat[:] = Ftau_real + 1j * Ftau_imag
+            Ftau_flat[:] = (Ftau_real + 1j * Ftau_imag).astype(Ftau_flat.dtype)
         else:
             fused_spreading_3d_real(
                 xmod.astype(np.float64),
